+++ modules/common/src/mail/mailer.ts
@@ -1,31 +1,27 @@
 import { ReadyToSendEmail } from './ready-to-send-email.js';
-import { Base64Ratchet, ErrorRatchet, Logger, RequireRatchet, StringRatchet } from '@bitblit/ratchet-common';
-import { SendRawEmailCommand, SendRawEmailCommandOutput, SendRawEmailRequest, SendRawEmailResponse, SESClient } from '@aws-sdk/client-ses';
 import { MailerConfig } from './mailer-config.js';
 import { ResolvedReadyToSendEmail } from './resolved-ready-to-send-email.js';
 import { EmailAttachment } from './email-attachment.js';
-import { DateTime } from 'luxon';
 import { MailerLike } from './mailer-like.js';
+import { RequireRatchet } from '../lang/require-ratchet';
+import { ErrorRatchet } from '../lang/error-ratchet';
+import { Logger } from '../logger/logger';
+import { StringRatchet } from '../lang/string-ratchet';
+import { Base64Ratchet } from '../lang/base64-ratchet';
+import { SendEmailResult } from './send-email-result';
+import { ArchiveEmailResult } from './archive-email-result';
 
 /**
- * Generic Mail Sender for AWS.
+ * Generic Mail Sender
  *
  * Params:
  * ses: AWS SES handler, properly configured
  * defaultSendingAddress:
  */
-export class Mailer implements MailerLike {
-  public static readonly EMAIL: RegExp = new RegExp('.+@.+\\.[a-z]+');
-
-  constructor(private ses: SESClient, private config: MailerConfig = {} as MailerConfig) {
-    RequireRatchet.notNullOrUndefined(this.ses);
-    if (!!config.archive && !config.archive.getDefaultBucket()) {
-      throw new Error('If archive specified, must set a default bucket');
-    }
-  }
-
-  public get sESClient(): SESClient {
-    return this.ses;
+export class Mailer<T, R> implements MailerLike<T, R> {
+  constructor(private config: MailerConfig<T, R>) {
+    RequireRatchet.notNullOrUndefined(config, 'config');
+    RequireRatchet.notNullOrUndefined(config.provider, 'config.provider');
   }
 
   public async fillEmailBody(
@@ -34,7 +30,7 @@ export class Mailer implements MailerLike {
     htmlTemplateName: string,
     txtTemplateName: string = null,
     layoutName: string = null,
-    partialNames: string[] = null
+    partialNames: string[] = null,
   ): Promise<ReadyToSendEmail> {
     RequireRatchet.notNullOrUndefined(htmlTemplateName);
     if (!this.config.templateRenderer) {
@@ -51,10 +47,10 @@ export class Mailer implements MailerLike {
     htmlTemplateName: string,
     txtTemplateName: string = null,
     layoutName: string = null,
-    partialNames: string[] = null
-  ): Promise<SendRawEmailResponse> {
+    partialNames: string[] = null,
+  ): Promise<SendEmailResult<T, R>> {
     const newVal: ReadyToSendEmail = await this.fillEmailBody(rts, context, htmlTemplateName, txtTemplateName, layoutName, partialNames);
-    const rval: SendRawEmailResponse = await this.sendEmail(newVal);
+    const rval: SendEmailResult<T, R> = await this.sendEmail(newVal);
     return rval;
   }
 
@@ -70,32 +66,24 @@ export class Mailer implements MailerLike {
     return rval;
   }
 
-  private async archiveEmailIfConfigured(rts: ResolvedReadyToSendEmail): Promise<boolean> {
-    let rval: boolean = false;
-    if (!!rts && !!this.config.archive && !rts.doNotArchive) {
+  private async archiveEmailIfConfigured(rts: ResolvedReadyToSendEmail, rawSendResult: T): Promise<ArchiveEmailResult<R>> {
+    let rval: ArchiveEmailResult<R> = null;
+    if (!!rts && !!this.config.provider.archiveEmail && !rts.doNotArchive) {
       Logger.debug('Archiving outbound email to : %j', rts.destinationAddresses);
-      let targetPath: string = StringRatchet.trimToEmpty(this.config.archivePrefix);
-      if (!targetPath.endsWith('/')) {
-        targetPath += '/';
-      }
-      const now: DateTime = DateTime.utc();
-      targetPath +=
-        'year=' +
-        now.toFormat('yyyy') +
-        '/month=' +
-        now.toFormat('MM') +
-        '/day=' +
-        now.toFormat('dd') +
-        '/hour=' +
-        now.toFormat('HH') +
-        '/' +
-        now.toFormat('mm_ss__SSS');
-      targetPath += '.json';
       try {
-        await this.config.archive.writeObjectToCacheFile(targetPath, rts);
-        rval = true;
+        const raw: R = await this.config.provider.archiveEmail(rts, rawSendResult);
+        rval = {
+          raw: raw,
+          error: null,
+          meta: {},
+        };
       } catch (err) {
-        Logger.warn('Failed to archive email %s %j : %s', targetPath, rts, err);
+        Logger.warn('Failed to archive email %j : %s', rts, err);
+        rval = {
+          raw: null,
+          error: ErrorRatchet.safeStringifyErr(err),
+          meta: {},
+        };
       }
     }
     return rval;
@@ -151,7 +139,7 @@ export class Mailer implements MailerLike {
                   ' was removed since it was ' +
                   a.base64Data.length +
                   ' bytes but max allowed is ' +
-                  this.config.maxAttachmentSizeInBase64Bytes
+                  this.config.maxAttachmentSizeInBase64Bytes,
               ),
             });
           }
@@ -161,10 +149,42 @@ export class Mailer implements MailerLike {
     }
   }
 
-  public async sendEmail(inRts: ReadyToSendEmail): Promise<SendRawEmailCommandOutput> {
+  public async sendEmail(inRts: ReadyToSendEmail): Promise<SendEmailResult<T, R>> {
     RequireRatchet.notNullOrUndefined(inRts, 'RTS must be defined');
     RequireRatchet.notNullOrUndefined(inRts.destinationAddresses, 'Destination addresses must be defined');
-    let rval: SendRawEmailCommandOutput = null;
+    let rval: SendEmailResult<T, R> = null;
+
+    const resolved: ResolvedReadyToSendEmail = await this.resolveReadyToSendEmail(inRts);
+    try {
+      const raw: T = await this.config.provider.sendEmail(resolved);
+      const archiveResult: ArchiveEmailResult<R> = await this.archiveEmailIfConfigured(resolved, raw);
+      rval = {
+        request: inRts,
+        resolved: resolved,
+        success: true,
+        rawResult: raw,
+        error: null,
+        meta: {},
+        archiveResults: archiveResult,
+      };
+    } catch (err) {
+      rval = {
+        request: inRts,
+        resolved: resolved,
+        success: false,
+        rawResult: null,
+        error: ErrorRatchet.safeStringifyErr(err),
+        meta: {},
+        archiveResults: null,
+      };
+    }
+
+    return rval;
+  }
+
+  public async resolveReadyToSendEmail(inRts: ReadyToSendEmail): Promise<ResolvedReadyToSendEmail> {
+    RequireRatchet.notNullOrUndefined(inRts, 'RTS must be defined');
+    RequireRatchet.notNullOrUndefined(inRts.destinationAddresses, 'Destination addresses must be defined');
 
     let toAddresses: string[] = this.filterEmailsToValid(inRts.destinationAddresses);
     const autoBcc: string[] = inRts.doNotAutoBcc ? [] : this.config.autoBccAddresses || [];
@@ -183,102 +203,12 @@ export class Mailer implements MailerLike {
     this.applyLimitsToBodySizesIfAnyInPlace(rts);
     this.applyLimitsToAttachmentSizesIfAnyInPlace(rts);
 
-    await this.archiveEmailIfConfigured(rts);
-
     if (rts.destinationAddresses.length === 0) {
       Logger.info('After cleaning email lists, no destination addresses left - not sending email');
-    } else {
-      const toLine: string = 'To: ' + rts.destinationAddresses.join(', ') + '\n';
-      const bccLine: string = !!rts.bccAddresses && rts.bccAddresses.length > 0 ? 'Bcc: ' + rts.bccAddresses.join(', ') + '\n' : '';
-
-      try {
-        const from: string = rts.fromAddress || this.config.defaultSendingAddress;
-        const boundary: string = 'NextPart';
-        const altBoundary: string = 'AltPart';
-        let rawMail: string = 'From: ' + from + '\n';
-        rawMail += toLine;
-        rawMail += bccLine;
-        rawMail += 'Subject: ' + rts.subject + '\n';
-        rawMail += 'MIME-Version: 1.0\n';
-        rawMail += 'Content-Type: multipart/mixed; boundary="' + boundary + '"\n';
-
-        rawMail += '\n\n--' + boundary + '\n';
-        rawMail += 'Content-Type: multipart/alternative; boundary="' + altBoundary + '"\n';
-        if (!!StringRatchet.trimToNull(rts.htmlMessage)) {
-          rawMail += '\n\n--' + altBoundary + '\n';
-          rawMail += 'Content-Type: text/html; charset="UTF-8"\n\n';
-          rawMail += rts.htmlMessage;
-        }
-        if (!!StringRatchet.trimToNull(rts.txtMessage)) {
-          rawMail += '\n\n--' + altBoundary + '\n';
-          rawMail += 'Content-Type: text/plain\n\n';
-          rawMail += rts.txtMessage;
-        }
-
-        rawMail += '\n\n--' + altBoundary + '--\n';
-
-        if (rts.attachments) {
-          rts.attachments.forEach((a) => {
-            rawMail += '\n\n--' + boundary + '\n';
-            rawMail += 'Content-Type: ' + a.contentType + '; name="' + a.filename + '"\n';
-            rawMail += 'Content-Transfer-Encoding: base64\n';
-            rawMail += 'Content-Disposition: attachment\n\n';
-            rawMail += a.base64Data.replace(/([^\0]{76})/g, '$1\n') + '\n\n';
-          });
-        }
-        rawMail += '\n\n--' + boundary + '--\n';
-
-        const params: SendRawEmailRequest = {
-          RawMessage: { Data: new TextEncoder().encode(rawMail) },
-        };
-
-        rval = await this.ses.send(new SendRawEmailCommand(params));
-      } catch (err) {
-        Logger.error('Error while processing email: %s', err, err);
-      }
     }
-    return rval;
+
+    rts.fromAddress = rts.fromAddress || this.config.defaultSendingAddress;
+
+    return rts;
   }
-
-  public static validEmail(email: string): boolean {
-    return email !== null && Mailer.EMAIL.test(email);
-  }
-
-  /*
-    public async sendEmail2(rts:ReadyToSendEmail): Promise<SendEmailResponse> {
-        let rval: SendEmailResponse = null;
-
-        try {
-            //const p: SendRawEmailRequest = {};
-
-            const params: SendEmailRequest = {
-                Destination: {
-                    ToAddresses: rts.destinationAddresses,
-                    BccAddresses: this.autoBccAddresses
-                },
-                Message: {
-                    Body: {
-                        Html: {
-                            Data: rts.htmlMessage
-                        },
-                        Text: {
-                            Data: rts.txtMessage
-                        }
-                    },
-                    Subject: {
-                        Data: rts.subject
-                    }
-                },
-                Source: rts.fromAddress || this.defaultSendingAddress
-            };
-
-            rval = await this.ses.sendEmail(params);
-
-            Logger.debug('Got send result : %j', rval);
-        } catch (err) {
-            Logger.error("Error while processing email: %s" ,err,err);
-        }
-        return rval;
-    }
-    */
 }